Span title across multiple grid columns
Previously, because the title was so large and occupied column 1, the
layout manager forced the width of column 1 to expand to accomodate the
size of the title in the specified font.

A side-effect of this choice was the all items in column 1 seemed to
consume the same amount of horizontal space as the title. A visual
consequence of this choice was a large amount of horizontal space
between our text box "label" and the text box itself.

To address this issue, we allowed the title to span three (auto width)
columns. Consequently, the default width of the name label was as large
as the text itself. Further, this causes the text box "label" and the
edit text box itself to be more "reasonably" aligned (no large empty
horizontal space - for now).